+++ src/NetworkDynamics.jl
@@ -106,6 +106,8 @@ export has_default, get_default, set_default!, delete_default!, set_defaults!, s
 export has_guess, get_guess, set_guess!, delete_guess!, strip_guesses!
 export has_init, get_init, set_init!, delete_init!, strip_inits!
 export has_bounds, get_bounds, set_bounds!, delete_bounds!, strip_bounds!
+export has_scope, get_scope, set_scope!, delete_scope!, strip_scopes!
+export VariableScope
 export has_graphelement, get_graphelement, set_graphelement!
 export get_initial_state, dump_initial_state, dump_state
 export has_callback, get_callbacks, set_callback!, add_callback!, delete_callbacks!
@@ -128,7 +130,9 @@ include("linear_analysis.jl")
 include("show.jl")
 
 const CHECK_COMPONENT = Ref(true)
-export chk_component
+# controls whether `chk_global_parameters` is run automatically on ODEProblem construction
+const CHECK_GLOBAL_PARAMETERS = Ref(true)
+export chk_component, chk_global_parameters
 include("doctor.jl")
 
 # additional utils, which depend on specific types beeing defined so they should be