+++ test/runtests.jl
@@ -50,6 +50,7 @@ BUILDKITE && @test CUDA.functional() # fail early in buildkite if cuda is not av
         @testfile "sparsity_test.jl"
 
         @testfile "MTK_test.jl"
+        @testfile "global_paramcheck.jl"
 
         # check on the precompile files
         @testfile "../src/precompile_workload.jl"